Экшен для стора рефреш
