+++ src/app/templates/store/actions/index.ts
@@ -1,0 +1,4 @@
+export { Refresh } from './refresh.action';
+export { ResetFilters } from './reset.action';
+export { ApplyFilters } from './apply-filters.action';
+export { Paginate } from './paginate.action';
