+++ src/app/templates/store/actions/refresh.action.ts
@@ -1,0 +1,6 @@
+import { Action } from '@ngrx/store';
+import { CollectionActionTypes } from '../types';
+
+export class Refresh implements Action {
+  public readonly type = CollectionActionTypes.REFRESH;
+}
