+++ src/app/templates/store/actions/reset.action.ts
@@ -1,5 +1,5 @@
 import { Action } from '@ngrx/store';
-import { CollectionActionTypes } from './types';
+import { CollectionActionTypes } from '../types';
 
 export class ResetFilters implements Action {
   public readonly type = CollectionActionTypes.RESET_FILTERS;
