+++ src/app/templates/store/reducer.ts
@@ -1,5 +1,5 @@
 import { CollectionActionsUnion, CollectionActionTypes } from './types';
-import { ITemplateQueryParams } from '../../models/template';
+import { ITemplateQueryParams } from '../models/template';
 
 export const INITIAL_FILTERS_STATE: ITemplateQueryParams = {
 	searchStr: '',
@@ -21,20 +21,27 @@ export function reducer(
 		}
 
 		case CollectionActionTypes.APPLY_FILTERS: {
-			state.selectedCategory = action.filters.selectedCategory;
-			state.searchStr = action.filters.searchStr;
-			state.sortBy = action.filters.sortBy;
+			state.selectedCategory = action.filters.selectedCategory
+				? action.filters.selectedCategory
+				: state.selectedCategory;
+			state.searchStr = action.filters.searchStr
+				? action.filters.searchStr
+				: state.searchStr;
+			state.sortBy = action.filters.sortBy
+				? action.filters.sortBy
+				: state.sortBy;
 			state.skip = 0;
 			return state;
 		}
 
 		case CollectionActionTypes.RESET_FILTERS: {
-      state.selectedCategory = INITIAL_FILTERS_STATE.selectedCategory;
-      state.searchStr = INITIAL_FILTERS_STATE.searchStr;
-      state.sortBy = INITIAL_FILTERS_STATE.sortBy;
+			state.selectedCategory = INITIAL_FILTERS_STATE.selectedCategory;
+			state.searchStr = INITIAL_FILTERS_STATE.searchStr;
+			state.sortBy = INITIAL_FILTERS_STATE.sortBy;
 			return state;
 		}
 
+    case CollectionActionTypes.REFRESH:
 		default: {
 			return state;
 		}
